src/server_flask/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
gevent==24.11.1
gevent-websocket==0.10.1
python-dotenv==1.1.0
openai==1.78.0
python-socketio==5.12.1
pydantic==2.11.3
PyJWT==2.8.0
bcrypt==4.1.2
psycopg[binary]>=3.2,<3.3
